+++ JSha256.h
@@ -1,10 +1,6 @@
 #ifndef _JSha256_H__
 #define _JSha256_H__
  
-#include<string.h>
-#include<stdio.h>
-#include<stdint.h>
-
 //储存SHA256加密数据
 struct Sha256Context{
 	char hash[32];//二进制数据
@@ -16,10 +12,10 @@ struct Sha256Context{
 typedef struct Sha256Context shacontext;
 
  //哈希加密函数
-void jsha_hash(const char *data, size_t len, unsigned char** outptr_ptr);
+void jsha_hash(const char *data,int len, unsigned char** outptr_ptr);
 
 //获取哈希加密存储结构
-void jsha_getobj(const char *data, size_t len,shacontext** contextptr_ptr);
+void jsha_getobj(const char *data,int len,shacontext** contextptr_ptr);
 //打印哈希数据存储结构
 void jsha_printobj(shacontext* context);
 char* jsha_getjson(shacontext* context);
